refactor: move PDF boilerplate to standard stack
- Reusable components and routes
