+++ app/components/quotes/QuotePdfTemplate.tsx
@@ -1,4 +1,5 @@
 import type { QuoteWithRelations } from "~/lib/quotes";
+import { formatCurrency, formatDate, commonPdfStyles } from "~/lib/pdf-utils";
 
 interface QuotePdfTemplateProps {
   quote: QuoteWithRelations;
@@ -6,23 +7,6 @@ interface QuotePdfTemplateProps {
 }
 
 export function QuotePdfTemplate({ quote, editable = false }: QuotePdfTemplateProps) {
-  const formatCurrency = (amount: number | string | null) => {
-    if (amount === null || amount === undefined) return "$0.00";
-    const num = typeof amount === "string" ? parseFloat(amount) : amount;
-    return new Intl.NumberFormat("en-US", {
-      style: "currency",
-      currency: "USD",
-    }).format(num);
-  };
-
-  const formatDate = (date: Date | null) => {
-    if (!date) return "TBD";
-    return new Intl.DateTimeFormat("en-US", {
-      year: "numeric",
-      month: "long",
-      day: "numeric",
-    }).format(new Date(date));
-  };
 
   const partsLineItems = (quote.lineItems || []).filter((item) => item.quotePartId !== null);
   const serviceLineItems = (quote.lineItems || []).filter((item) => item.quotePartId === null);
@@ -43,22 +27,7 @@ export function QuotePdfTemplate({ quote, editable = false }: QuotePdfTemplatePr
     <div className="quote-pdf-container">
       <style>
         {`
-          .quote-pdf-container {
-            font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-            background-color: transparent;
-            padding: 20px;
-            color: #2c3e50;
-            line-height: 1.6;
-          }
-
-          .quote-container {
-            max-width: 1400px;
-            margin: 0 auto;
-            background-color: white;
-            box-shadow: 0 2px 20px rgba(0, 0, 0, 0.08);
-            display: flex;
-            flex-direction: column;
-          }
+          ${commonPdfStyles}
 
           .content-wrapper {
             flex: 1;
@@ -232,50 +201,6 @@ export function QuotePdfTemplate({ quote, editable = false }: QuotePdfTemplatePr
             color: #c41e3a;
           }
 
-          table {
-            width: 100%;
-            border-collapse: collapse;
-            margin-bottom: 12px;
-            border: 1px solid #dee2e6;
-          }
-
-          thead {
-            background-color: #f8f9fa;
-          }
-
-          th {
-            padding: 8px 12px;
-            text-align: left;
-            font-weight: 700;
-            color: #2c3e50;
-            font-size: 12px;
-            text-transform: uppercase;
-            letter-spacing: 0.5px;
-            border-bottom: 2px solid #dee2e6;
-            border-right: 1px solid #e9ecef;
-          }
-
-          th:last-child {
-            border-right: none;
-          }
-
-          td {
-            padding: 8px 12px;
-            color: #495057;
-            font-size: 14px;
-            border-bottom: 1px solid #e9ecef;
-            border-right: 1px solid #e9ecef;
-            background: white;
-          }
-
-          td:last-child {
-            border-right: none;
-          }
-
-          tbody tr:last-child td {
-            border-bottom: none;
-          }
-
           .financial-section {
             padding: 20px 40px 20px;
           }
@@ -352,52 +277,10 @@ export function QuotePdfTemplate({ quote, editable = false }: QuotePdfTemplatePr
           .footer strong {
             color: #c41e3a;
           }
-
-          .editable {
-            transition: background-color 0.15s ease;
-            border-radius: 3px;
-            padding: 4px 6px;
-            margin: -4px -6px;
-            min-width: 30px;
-            display: inline-block;
-          }
-
-          .editable:hover {
-            background-color: #fff3cd;
-            cursor: text;
-            outline: 2px dashed #ffc107;
-            outline-offset: 2px;
-          }
-
-          .editable:focus {
-            background-color: #fff8dc;
-            outline: 2px solid #ffc107;
-            outline-offset: 2px;
-          }
-
-          @media print {
-            .quote-pdf-container {
-              background-color: white;
-              padding: 0;
-              margin: 0;
-            }
-
-            .quote-container {
-              box-shadow: none;
-              max-width: 100%;
-              min-height: 11in;
-            }
-
-            .editable:hover,
-            .editable:focus {
-              background-color: transparent;
-              outline: none;
-            }
-          }
         `}
       </style>
 
-      <div className="quote-container">
+      <div className="document-container">
         <div className="content-wrapper">
           {/* Header */}
           <div className="header">
